Introduce 'untested' category for audit tools
Without an 'untested' category you would have to remove
all results if there weren't any, which is
an additional task for anyone adding new issues.
It would also make it more difficult to add results back later,
as the whole array would need to be copied again.

With this category it's easier to ignore the section
as well as add test results later.